Refactored coordination manager supporting responsibilities to multiple separate types
