+++ src/AI4E.Coordination/CoordinationManager.cs
@@ -38,15 +38,11 @@
 using System;
 using System.Collections.Generic;
 using System.Diagnostics;
-using System.IO;
 using System.Linq;
-using System.Net.Sockets;
-using System.Text;
 using System.Threading;
 using System.Threading.Tasks;
 using AI4E.Async;
 using AI4E.Internal;
-using AI4E.Processing;
 using AI4E.Remoting;
 using Microsoft.Extensions.Logging;
 using static System.Diagnostics.Debug;
@@ -78,15 +74,14 @@ namespace AI4E.Coordination
         private readonly ILogger<CoordinationManager<TAddress>> _logger;
 
         private readonly CoordinationEntryCache _cache;
+        private readonly ICoordinationWaitManager _waitManager;
+        private readonly ICoordinationLockManager _lockManager;
+        private readonly ICoordinationCacheManager _cacheManager;
+        private readonly ICoordinationExchangeManager _exchangeManager;
 
-        private readonly AsyncWaitDirectory<(Session session, CoordinationEntryPath path)> _readLockWaitDirectory;
-        private readonly AsyncWaitDirectory<(Session session, CoordinationEntryPath path)> _writeLockWaitDirectory;
-
-        private readonly IAsyncProcess _updateSessionProcess;
-        private readonly IAsyncProcess _sessionCleanupProcess;
-        private readonly IAsyncProcess _receiveProcess;
-        private readonly AsyncInitializationHelper<(Session session, IPhysicalEndPoint<TAddress> physicalEndPoint)> _initializationHelper;
+        private readonly AsyncInitializationHelper<Session> _initializationHelper;
         private readonly AsyncDisposeHelper _disposeHelper;
+        private readonly SessionManagement _sessionManagement;
 
         #endregion
 
@@ -132,337 +127,31 @@ namespace AI4E.Coordination
             _logger = logger;
 
             _cache = new CoordinationEntryCache();
-            _readLockWaitDirectory = new AsyncWaitDirectory<(Session session, CoordinationEntryPath path)>();
-            _writeLockWaitDirectory = new AsyncWaitDirectory<(Session session, CoordinationEntryPath path)>();
+            _exchangeManager = new ExchangeManager(this, _sessionManager, Provider.Create(() => _waitManager), Provider.Create(() => _cacheManager), _endPointMultiplexer, _addressConversion, _logger);
+            _waitManager = new WaitManager(this, _storage, _storedEntryManager, _sessionManager, _exchangeManager, _logger);
+            _lockManager = new LockManager(this, _cache, _storage, _storedEntryManager, _waitManager, _exchangeManager, _logger);
+            _cacheManager = new CacheManager(_cache, _lockManager);
 
-            _updateSessionProcess = new AsyncProcess(UpdateSessionProcess);
-            _sessionCleanupProcess = new AsyncProcess(SessionCleanupProcess);
-            _receiveProcess = new AsyncProcess(ReceiveProcess);
-            _initializationHelper = new AsyncInitializationHelper<(Session session, IPhysicalEndPoint<TAddress> physicalEndPoint)>(InitializeInternalAsync);
+            _initializationHelper = new AsyncInitializationHelper<Session>(InitializeInternalAsync);
             _disposeHelper = new AsyncDisposeHelper(DisposeInternalAsync);
-        }
-
-        #endregion
-
-        #region RX/TX
-
-        private async Task ReceiveProcess(CancellationToken cancellation)
-        {
-            while (cancellation.ThrowOrContinue())
-            {
-                try
-                {
-                    var message = await ReceiveMessageAsync(cancellation);
-                    var (messageType, path, session) = DecodeMessage(message);
-
-                    Task.Run(() => HandleMessageAsync(message, messageType, path, session, cancellation)).HandleExceptions();
-                }
-                catch (OperationCanceledException) when (cancellation.IsCancellationRequested) { throw; }
-                catch (Exception exc)
-                {
-                    _logger?.LogWarning(exc, $"[{await GetSessionAsync()}] Failure while decoding received message.");
-                }
-            }
-        }
-
-        private async Task HandleMessageAsync(IMessage message, MessageType messageType, CoordinationEntryPath path, Session session, CancellationToken cancellation)
-        {
-            switch (messageType)
-            {
-                case MessageType.InvalidateCacheEntry:
-                    if (session != await GetSessionAsync(cancellation))
-                    {
-                        _logger?.LogWarning($"[{await GetSessionAsync()}] Received invalidate message for session that is not present.");
-                    }
-                    else
-                    {
-                        await InvalidateCacheEntryAsync(path, cancellation);
-                    }
-                    break;
-
-                case MessageType.ReleasedReadLock:
-                    _readLockWaitDirectory.Notify((session, path));
-
-                    break;
-
-                case MessageType.ReleasedWriteLock:
-                    _writeLockWaitDirectory.Notify((session, path));
-
-                    break;
-
-                case MessageType.Unknown:
-                default:
-                    _logger?.LogWarning($"[{await GetSessionAsync()}] Received invalid message or message with unknown message type.");
-                    break;
-            }
-        }
-
-        private async Task<IMessage> ReceiveMessageAsync(CancellationToken cancellation)
-        {
-            var physicalEndPoint = await GetPhysicalEndPointAsync(cancellation);
-
-            return await physicalEndPoint.ReceiveAsync(cancellation);
-        }
-
-        private async Task InvalidateCacheEntryAsync(CoordinationEntryPath path, Session session, CancellationToken cancellation)
-        {
-            if (session == await GetSessionAsync(cancellation))
-            {
-                await InvalidateCacheEntryAsync(path, cancellation);
-            }
-            else
-            {
-                // The session is the read-lock owner (It caches the entry currently)
-                var message = EncodeMessage(MessageType.InvalidateCacheEntry, path, session);
-
-                await SendMessageAsync(session, message, cancellation);
-            }
-        }
-
-        private async Task NotifyReadLockReleasedAsync(CoordinationEntryPath path, CancellationToken cancellation = default)
-        {
-            var sessions = await _sessionManager.GetSessionsAsync(cancellation);
-            var localSession = await GetSessionAsync(cancellation);
-
-            foreach (var session in sessions)
-            {
-                if (session == localSession)
-                {
-                    _readLockWaitDirectory.Notify((localSession, path));
-
-                    continue;
-                }
-
-                // The session is the former read-lock owner.
-                var message = EncodeMessage(MessageType.ReleasedReadLock, path, localSession);
-
-                await SendMessageAsync(session, message, cancellation);
-            }
-        }
-
-        private async Task NotifyWriteLockReleasedAsync(CoordinationEntryPath path, CancellationToken cancellation = default)
-        {
-            var sessions = await _sessionManager.GetSessionsAsync(cancellation);
-            var localSession = await GetSessionAsync(cancellation);
-
-            foreach (var session in sessions)
-            {
-                if (session == localSession)
-                {
-                    _writeLockWaitDirectory.Notify((localSession, path));
-
-                    continue;
-                }
-
-                // The session is the former write-lock owner.
-                var message = EncodeMessage(MessageType.ReleasedWriteLock, path, localSession);
-
-                await SendMessageAsync(session, message, cancellation);
-            }
-        }
-
-        private async Task SendMessageAsync(Session session, Message message, CancellationToken cancellation)
-        {
-            var remoteAddress = _addressConversion.DeserializeAddress(session.PhysicalAddress.ToArray()); // TODO: This will copy everything to a new aray
-
-            Assert(remoteAddress != null);
-
-            var physicalEndPoint = GetSessionEndPoint(session);
-
-            try
-            {
-                await physicalEndPoint.SendAsync(message, remoteAddress, cancellation);
-            }
-            catch (SocketException) { }
-            catch (IOException) { } // The remote session terminated or we just cannot transmit to it.
-
-        }
-
-        private IPhysicalEndPoint<TAddress> GetSessionEndPoint(Session session)
-        {
-            Assert(session != null);
-
-            var multiplexName = GetMultiplexEndPointName(session);
-
-            var result = _endPointMultiplexer.GetPhysicalEndPoint(multiplexName);
-
-            Assert(result != null);
-
-            return result;
-        }
-
-        private static string GetMultiplexEndPointName(Session session)
-        {
-            return "coord/" + session.ToCompactString();
-        }
-
-        private (MessageType messageType, CoordinationEntryPath path, Session session) DecodeMessage(IMessage message)
-        {
-            Assert(message != null);
-
-            using (var frameStream = message.PopFrame().OpenStream())
-            using (var binaryReader = new BinaryReader(frameStream))
-            {
-                var messageType = (MessageType)binaryReader.ReadByte();
-
-                var escapedPath = binaryReader.ReadUtf8();
-                var path = CoordinationEntryPath.FromEscapedPath(escapedPath);
-
-                var sessionLength = binaryReader.ReadInt32();
-                var sessionBytes = binaryReader.ReadBytes(sessionLength);
-                var session = Session.FromChars(Encoding.UTF8.GetString(sessionBytes).AsSpan());
-
-                return (messageType, path, session);
-            }
-        }
-
-        private Message EncodeMessage(MessageType messageType, CoordinationEntryPath path, Session session)
-        {
-            var message = new Message();
-
-            EncodeMessage(message, messageType, path, session);
-
-            return message;
-        }
-
-        private void EncodeMessage(IMessage message, MessageType messageType, CoordinationEntryPath path, Session session)
-        {
-            Assert(message != null);
-            // Modify if other message types are added
-            Assert(messageType >= MessageType.InvalidateCacheEntry && messageType <= MessageType.ReleasedWriteLock);
-
-            using (var frameStream = message.PushFrame().OpenStream())
-            using (var binaryWriter = new BinaryWriter(frameStream))
-            {
-                binaryWriter.Write((byte)messageType);
-
-                binaryWriter.WriteUtf8(path.EscapedPath.Span);
-
-                var sessionBytes = Encoding.UTF8.GetBytes(session.ToCompactString());
-                binaryWriter.Write(sessionBytes.Length);
-                binaryWriter.Write(sessionBytes);
-            }
-        }
-
-        private enum MessageType : byte
-        {
-            Unknown = 0,
-            InvalidateCacheEntry = 1,
-            ReleasedReadLock = 2,
-            ReleasedWriteLock = 3
-        }
-
-        #endregion
-
-        #region SessionManagement
-
-        private async Task SessionCleanupProcess(CancellationToken cancellation)
-        {
-            var session = await GetSessionAsync(cancellation);
-
-            _logger?.LogTrace($"[{await GetSessionAsync()}] Started session cleanup process.");
-
-            while (cancellation.ThrowOrContinue())
-            {
-                try
-                {
-                    var terminated = await _sessionManager.WaitForTerminationAsync(cancellation);
-
-                    Assert(terminated != null);
-
-                    // Our session is terminated or
-                    // There are no session in the session manager. => Our session must be terminated.
-                    if (terminated == session)
-                    {
-                        Dispose();
-                    }
-                    else
-                    {
-                        await CleanupSessionAsync(terminated, cancellation);
-                    }
-                }
-                catch (OperationCanceledException) when (cancellation.IsCancellationRequested) { return; }
-                catch (OperationCanceledException) when (_disposeHelper.IsDisposed)
-                {
-                    return; // TODO: https://github.com/AI4E/AI4E/issues/37
-                }
-                catch (Exception exc)
-                {
-                    _logger?.LogWarning(exc, $"[{await GetSessionAsync()}] Failure while cleaning up terminated sessions.");
-                }
-            }
-        }
-
-        private async Task UpdateSessionProcess(CancellationToken cancellation)
-        {
-            var session = await GetSessionAsync(cancellation);
-
-            while (cancellation.ThrowOrContinue())
-            {
-                try
-                {
-                    await UpdateSessionAsync(session, cancellation);
-                }
-                catch (OperationCanceledException) when (cancellation.IsCancellationRequested) { throw; }
-                catch (Exception exc)
-                {
-                    _logger?.LogWarning(exc, $"[{await GetSessionAsync()}] Failure while updating session {session}.");
-                }
-            }
-        }
-
-        private async Task UpdateSessionAsync(Session session, CancellationToken cancellation)
-        {
-            Assert(session != null);
-
-            var leaseEnd = _dateTimeProvider.GetCurrentTime() + _leaseLength;
-
-            try
-            {
-                await _sessionManager.UpdateSessionAsync(session, leaseEnd, cancellation);
-
-                await Task.Delay(_leaseLengthHalf);
-            }
-            catch (SessionTerminatedException)
-            {
-                Dispose();
-            }
-        }
-
-        private async Task CleanupSessionAsync(Session session, CancellationToken cancellation)
-        {
-            _logger?.LogInformation($"[{await GetSessionAsync()}] Cleaning up session '{session}'.");
-
-            var entries = await _sessionManager.GetEntriesAsync(session, cancellation);
-
-            await Task.WhenAll(entries.Select(async entry =>
-            {
-                await DeleteAsync(entry, version: default, recursive: false, cancellation);
-                await _sessionManager.RemoveSessionEntryAsync(session, entry, cancellation);
-            }));
-
-            await _sessionManager.EndSessionAsync(session, cancellation);
+            _sessionManagement = new SessionManagement(this, _sessionManager, _dateTimeProvider, _logger);
         }
 
         #endregion
 
         #region Initialization
 
-        public async ValueTask<Session> GetSessionAsync(CancellationToken cancellation = default)
+        public ValueTask<Session> GetSessionAsync(CancellationToken cancellation = default)
         {
-            var (session, _) = await _initializationHelper.Initialization.WithCancellation(cancellation);
+            if (!cancellation.CanBeCanceled || _initializationHelper.Initialization.IsCompleted)
+            {
+                return new ValueTask<Session>(_initializationHelper.Initialization);
+            }
 
-            return session;
+            return new ValueTask<Session>(_initializationHelper.Initialization.WithCancellation(cancellation));
         }
 
-        private async Task<IPhysicalEndPoint<TAddress>> GetPhysicalEndPointAsync(CancellationToken cancellation = default)
-        {
-            var (_, physicalEndPoint) = await _initializationHelper.Initialization.WithCancellation(cancellation);
-
-            return physicalEndPoint;
-        }
-
-        private async Task<(Session session, IPhysicalEndPoint<TAddress> physicalEndPoint)> InitializeInternalAsync(CancellationToken cancellation)
+        private async Task<Session> InitializeInternalAsync(CancellationToken cancellation)
         {
             Session session;
 
@@ -474,33 +163,9 @@ namespace AI4E.Coordination
             }
             while (!await _sessionManager.TryBeginSessionAsync(session, leaseEnd: _dateTimeProvider.GetCurrentTime() + _leaseLength, cancellation));
 
-            try
-            {
-                var physicalEndPoint = GetSessionEndPoint(session);
+            _logger?.LogInformation($"[{session}] Initialized coordination manager.");
 
-                try
-                {
-                    await _updateSessionProcess.StartAsync(cancellation);
-                    await _sessionCleanupProcess.StartAsync(cancellation);
-                    await _receiveProcess.StartAsync(cancellation);
-                }
-                catch
-                {
-                    await physicalEndPoint.DisposeIfDisposableAsync().HandleExceptionsAsync(_logger);
-
-                    throw;
-                }
-
-                _logger?.LogInformation($"[{session}] Initialized coordination manager.");
-
-                return (session, physicalEndPoint);
-            }
-            catch
-            {
-                await _sessionManager.EndSessionAsync(session).HandleExceptionsAsync(_logger);
-
-                throw;
-            }
+            return session;
         }
 
         #endregion
@@ -521,7 +186,9 @@ namespace AI4E.Coordination
 
         private async Task DisposeInternalAsync()
         {
-            var (success, (session, physicalEndPoint)) = await _initializationHelper.CancelAsync().HandleExceptionsAsync(_logger);
+            var (success, session) = await _initializationHelper.CancelAsync().HandleExceptionsAsync(_logger);
+
+            _sessionManagement.Dispose();
 
             if (success)
             {
@@ -532,15 +199,9 @@ namespace AI4E.Coordination
                 _logger?.LogInformation($"Disposing coordination manager.");
             }
 
-            await Task.WhenAll(_sessionCleanupProcess.TerminateAsync().HandleExceptionsAsync(_logger),
-                               _updateSessionProcess.TerminateAsync().HandleExceptionsAsync(_logger),
-                               _receiveProcess.TerminateAsync().HandleExceptionsAsync(_logger));
-
             if (success)
             {
-                await Task.WhenAll(_sessionManager.EndSessionAsync(session).HandleExceptionsAsync(_logger),
-                                   physicalEndPoint.DisposeIfDisposableAsync().HandleExceptionsAsync(_logger));
-
+                await _sessionManager.EndSessionAsync(session).HandleExceptionsAsync(_logger);
             }
         }
 
@@ -560,7 +221,7 @@ namespace AI4E.Coordination
             return new Entry(this, entry);
         }
 
-        public async ValueTask<IStoredEntry> GetEntryAsync(CoordinationEntryPath path, CancellationToken cancellation)
+        private async ValueTask<IStoredEntry> GetEntryAsync(CoordinationEntryPath path, CancellationToken cancellation)
         {
             // First try to load the entry from the cache.
             var cacheEntry = _cache.GetEntry(path);
@@ -571,7 +232,7 @@ namespace AI4E.Coordination
                 return cacheEntry.Entry;
             }
 
-            var result = await AcquireReadLockAsync(path, cancellation);
+            var result = await _lockManager.AcquireReadLockAsync(path, cancellation);
 
             if (result == null)
             {
@@ -598,7 +259,7 @@ namespace AI4E.Coordination
             }
             catch
             {
-                await ReleaseReadLockAsync(path, cancellation);
+                await _lockManager.ReleaseReadLockAsync(path, cancellation);
                 throw;
             }
 
@@ -647,10 +308,7 @@ namespace AI4E.Coordination
                 }
 
                 Assert(entry != null);
-
-                await AddToCacheAsync(entry,
-                                      comparandVersion: default,
-                                      cancellation);
+                await _cacheManager.AddToCacheAsync(entry, cancellation);
 
                 return new Entry(this, entry);
             }
@@ -692,10 +350,7 @@ namespace AI4E.Coordination
                                                               combinedCancellationSource.Token);
 
                 Assert(entry != null);
-
-                await AddToCacheAsync(entry,
-                                      comparandVersion: default,
-                                      cancellation);
+                await _cacheManager.AddToCacheAsync(entry, cancellation);
 
                 return new Entry(this, entry);
             }
@@ -753,7 +408,7 @@ namespace AI4E.Coordination
                 finally
                 {
                     Assert(parent != null);
-                    await ReleaseWriteLockAsync(parent, cancellation);
+                    await _lockManager.ReleaseWriteLockAsync(parent, cancellation);
                 }
             }
             else
@@ -783,7 +438,7 @@ namespace AI4E.Coordination
         {
             IStoredEntry result;
 
-            while ((result = await AcquireWriteLockAsync(path, cancellation)) == null)
+            while ((result = await _lockManager.AcquireWriteLockAsync(path, cancellation)) == null)
             {
                 await TryCreateInternalAsync(path, ReadOnlyMemory<byte>.Empty, modes: default, session, cancellation);
             }
@@ -801,7 +456,7 @@ namespace AI4E.Coordination
         {
             // Lock the local-lock first.
             // Create will lock the write-lock for us.
-            await AcquireLocalLockAsync(path, cancellation);
+            await _lockManager.AcquireLocalLockAsync(path, cancellation);
 
             var entry = _storedEntryManager.Create(path, session, (modes & EntryCreationModes.Ephemeral) == EntryCreationModes.Ephemeral, value.Span);
             var comparand = await _storage.UpdateEntryAsync(entry, comparand: null, cancellation);
@@ -825,11 +480,11 @@ namespace AI4E.Coordination
                 // There is already an entry present => Release the local lock only.
                 if (comparand != null)
                 {
-                    await ReleaseLocalLockAsync(path, cancellation);
+                    await _lockManager.ReleaseLocalLockAsync(path, cancellation);
                 }
                 else
                 {
-                    await ReleaseWriteLockAsync(entry, cancellation);
+                    await _lockManager.ReleaseWriteLockAsync(entry, cancellation);
                 }
             }
 
@@ -875,7 +530,7 @@ namespace AI4E.Coordination
                 // There is no parent entry.
                 if (path.IsRoot)
                 {
-                    var entry = await AcquireWriteLockAsync(path, combinedCancellationSource.Token);
+                    var entry = await _lockManager.AcquireWriteLockAsync(path, combinedCancellationSource.Token);
                     try
                     {
                         return await DeleteInternalAsync(entry, session, version, recursive, combinedCancellationSource.Token);
@@ -883,11 +538,11 @@ namespace AI4E.Coordination
                     finally
                     {
                         Assert(entry != null);
-                        await ReleaseWriteLockAsync(entry, combinedCancellationSource.Token);
+                        await _lockManager.ReleaseWriteLockAsync(entry, combinedCancellationSource.Token);
                     }
                 }
 
-                var parent = await AcquireWriteLockAsync(path.GetParentPath(), combinedCancellationSource.Token);
+                var parent = await _lockManager.AcquireWriteLockAsync(path.GetParentPath(), combinedCancellationSource.Token);
 
                 try
                 {
@@ -897,7 +552,7 @@ namespace AI4E.Coordination
                         return 0;
                     }
 
-                    var entry = await AcquireWriteLockAsync(path, combinedCancellationSource.Token);
+                    var entry = await _lockManager.AcquireWriteLockAsync(path, combinedCancellationSource.Token);
 
                     if (entry == null)
                     {
@@ -929,13 +584,13 @@ namespace AI4E.Coordination
                     finally
                     {
                         Assert(entry != null);
-                        await ReleaseWriteLockAsync(entry, combinedCancellationSource.Token);
+                        await _lockManager.ReleaseWriteLockAsync(entry, combinedCancellationSource.Token);
                     }
                 }
                 finally
                 {
                     Assert(parent != null);
-                    await ReleaseWriteLockAsync(parent, combinedCancellationSource.Token);
+                    await _lockManager.ReleaseWriteLockAsync(parent, combinedCancellationSource.Token);
                 }
             }
         }
@@ -999,7 +654,7 @@ namespace AI4E.Coordination
 
                 // First load the child entry.
                 var childPath = entry.Path.GetChildPath(childName);
-                var child = await AcquireWriteLockAsync(childPath, cancellation);
+                var child = await _lockManager.AcquireWriteLockAsync(childPath, cancellation);
 
                 // The child-names collection is not guaranteed to be strongly consistent.
                 if (child == null)
@@ -1024,7 +679,7 @@ namespace AI4E.Coordination
                 finally
                 {
                     Assert(child != null);
-                    await ReleaseWriteLockAsync(child, cancellation);
+                    await _lockManager.ReleaseWriteLockAsync(child, cancellation);
                 }
 
                 // As we did not specify a version, the call must succeed.
@@ -1070,7 +725,7 @@ namespace AI4E.Coordination
 
                 await _initializationHelper.Initialization.WithCancellation(combinedCancellationSource.Token);
 
-                var entry = await AcquireWriteLockAsync(path, combinedCancellationSource.Token);
+                var entry = await _lockManager.AcquireWriteLockAsync(path, combinedCancellationSource.Token);
 
                 if (entry == null)
                 {
@@ -1090,90 +745,12 @@ namespace AI4E.Coordination
                 finally
                 {
                     Assert(entry != null);
-                    await ReleaseWriteLockAsync(entry, combinedCancellationSource.Token);
+                    await _lockManager.ReleaseWriteLockAsync(entry, combinedCancellationSource.Token);
                 }
             }
         }
 
         #endregion
-
-        #region Cache
-
-        private bool TryGetCacheEntry(CoordinationEntryPath path, out CacheEntry cacheEntry)
-        {
-            return _cache.TryGetEntry(path, out cacheEntry);
-        }
-
-        [Obsolete]
-        private void UpdateCacheEntry(IStoredEntry entry, int comparandVersion)
-        {
-            Assert(entry != null);
-
-            _cache.UpdateEntry(entry, comparandVersion);
-        }
-
-        private void UpdateCacheEntry(CacheEntry cacheEntry, IStoredEntry entry/*, int comparandVersion*/)
-        {
-            Assert(cacheEntry != null);
-            Assert(entry != null);
-
-            _cache.UpdateEntry(cacheEntry, entry);
-        }
-
-        [Obsolete]
-        private async Task<IStoredEntry> AddToCacheAsync(IStoredEntry entry, int comparandVersion, CancellationToken cancellation)
-        {
-            Assert(entry != null);
-
-            var path = entry.Path;
-
-            try
-            {
-                entry = await InternalAcquireReadLockAsync(entry, cancellation);
-
-                if (entry == null)
-                {
-                    await UpdateParentOfDeletedEntry(path, cancellation);
-                }
-
-                _cache.UpdateEntry(entry, comparandVersion);
-                return entry;
-            }
-            catch
-            {
-                await ReleaseReadLockAsync(path, cancellation);
-
-                throw;
-            }
-        }
-
-        private async Task<IStoredEntry> AddToCacheAsync(CacheEntry cacheEntry, IStoredEntry entry, CancellationToken cancellation)
-        {
-            Assert(cacheEntry != null);
-            Assert(entry != null);
-
-            var path = entry.Path;
-
-            try
-            {
-                entry = await InternalAcquireReadLockAsync(entry, cancellation);
-
-                if (entry == null)
-                {
-                    await UpdateParentOfDeletedEntry(path, cancellation);
-                }
-
-                _cache.UpdateEntry(cacheEntry, entry);
-
-                return entry;
-            }
-            catch
-            {
-                await ReleaseReadLockAsync(path, cancellation);
-
-                throw;
-            }
-        }
 
         private async Task UpdateParentOfDeletedEntry(CoordinationEntryPath path, CancellationToken cancellation)
         {
@@ -1189,7 +766,7 @@ namespace AI4E.Coordination
                     var session = await GetSessionAsync(cancellation);
 
                     // TODO: Can this be further optimized? We already loaded parent.
-                    parent = await AcquireWriteLockAsync(path, cancellation);
+                    parent = await _lockManager.AcquireWriteLockAsync(path, cancellation);
 
                     if (parent == null)
                     {
@@ -1209,644 +786,11 @@ namespace AI4E.Coordination
                     finally
                     {
                         Assert(parent != null);
-                        await ReleaseWriteLockAsync(parent, cancellation);
+                        await _lockManager.ReleaseWriteLockAsync(parent, cancellation);
                     }
                 }
             }
         }
-
-        private async Task InvalidateCacheEntryAsync(CoordinationEntryPath path, CancellationToken cancellation)
-        {
-            if (path == null)
-                throw new ArgumentNullException(nameof(path));
-
-            var session = await GetSessionAsync(cancellation);
-
-            _cache.InvalidateEntry(path);
-
-            var combinedCancellationSource = CancellationTokenSource.CreateLinkedTokenSource(cancellation, _disposeHelper.DisposalRequested);
-
-            _logger?.LogTrace($"[{await GetSessionAsync()}] Invalidating cache entry '{path.EscapedPath.ConvertToString()}'.");
-
-            using (await _disposeHelper.ProhibitDisposalAsync(cancellation))
-            {
-                if (_disposeHelper.IsDisposed)
-                    throw new ObjectDisposedException(GetType().FullName);
-
-                await _initializationHelper.Initialization.WithCancellation(combinedCancellationSource.Token);
-
-                var entry = await _storage.GetEntryAsync(path, combinedCancellationSource.Token);
-
-                await ReleaseReadLockAsync(path, cancellation);
-            }
-        }
-
-        #endregion
-
-        #region Locking
-
-        private async Task AcquireLocalLockAsync(CoordinationEntryPath path, CancellationToken cancellation)
-        {
-            var cacheEntry = _cache.GetEntry(path);
-            await cacheEntry.LocalLock.WaitAsync(cancellation);
-        }
-
-        private Task ReleaseLocalLockAsync(CoordinationEntryPath path, CancellationToken cancellation)
-        {
-            var cacheEntry = _cache.GetEntry(path);
-
-            Assert(cacheEntry != null);
-            Assert(cacheEntry.LocalLock.CurrentCount == 0);
-
-            cacheEntry.LocalLock.Release();
-            return Task.CompletedTask;
-        }
-
-        // Acquired a read lock for the entry with the specified path and returns the entry.
-        // If the result is null, the entry does not exist and no lock is allocated.
-        private async Task<IStoredEntry> AcquireWriteLockAsync(CoordinationEntryPath path, CancellationToken cancellation)
-        {
-            var cacheEntry = _cache.GetEntry(path);
-
-            // Enter local lock
-            await cacheEntry.LocalLock.WaitAsync(cancellation);
-
-            try
-            {
-                var entry = cacheEntry.IsValid ? cacheEntry.Entry : await _storage.GetEntryAsync(path, cancellation);
-
-                // Enter global lock
-                var result = await InternalAcquireWriteLockAsync(entry, cancellation);
-
-                // Release the local lock if the entry is null.
-                if (result == null)
-                {
-                    Assert(cacheEntry.LocalLock.CurrentCount == 0);
-                    cacheEntry.LocalLock.Release();
-                }
-
-                return result;
-            }
-            catch
-            {
-                Assert(cacheEntry.LocalLock.CurrentCount == 0);
-                // Release local lock on failure
-                cacheEntry.LocalLock.Release();
-                throw;
-            }
-        }
-
-        // Releases the write lock for the specified entry and returns the updated entry.
-        // If the current session does not own the write-lock for the entry (f.e. if it is deleted), 
-        // this method only releases the local lock but is a no-op otherwise.
-        private Task<IStoredEntry> ReleaseWriteLockAsync(IStoredEntry entry, CancellationToken cancellation)
-        {
-            Assert(entry != null);
-
-            return ReleaseWriteLockAsync(entry).WithCancellation(cancellation);
-        }
-
-        private async Task<IStoredEntry> ReleaseWriteLockAsync(IStoredEntry entry)
-        {
-            Assert(entry != null);
-
-            try
-            {
-                var result = await InternalReleaseWriteLockAsync(entry);
-                await ReleaseLocalLockAsync(entry.Path, cancellation: default);
-                return result;
-            }
-            catch (SessionTerminatedException) { throw; }
-            catch
-            {
-                await DisposeAsync();
-                throw;
-            }
-        }
-
-        private async Task<IStoredEntry> InternalAcquireWriteLockAsync(IStoredEntry entry, CancellationToken cancellation)
-        {
-            Stopwatch watch = null;
-            if (_logger?.IsEnabled(LogLevel.Trace) ?? false)
-            {
-                watch = new Stopwatch();
-                watch.Start();
-            }
-
-            var session = await GetSessionAsync(cancellation);
-
-            if (entry != null)
-            {
-                _logger?.LogTrace($"[{await GetSessionAsync(cancellation)}] Acquiring write-lock for entry '{entry.Path}'.");
-            }
-
-            // TODO: Check whether session is still alive
-
-            // We have to perform the operation in a CAS-like loop because concurrency is controlled via a version number only.
-            IStoredEntry start, desired;
-            do
-            {
-                // We wait till we are able to lock. This means we need to wait till the write lock is free.
-                start = await WaitForWriteLockReleaseAsync(entry, allowWriteLock: false, cancellation);
-
-                // The entry was deleted (concurrently).
-                if (start == null)
-                {
-                    return null;
-                }
-
-                Assert(start.WriteLock == null);
-
-                // Actually try to lock the entry.
-                // Do not use UpdateEntryAsync as this method assumes that we already own the write-lock.
-                desired = _storedEntryManager.AcquireWriteLock(start, session);
-                entry = await _storage.UpdateEntryAsync(desired, start, cancellation);
-            }
-            while (entry != start);
-
-            // If we reached this point, we own the write lock.
-            entry = desired;
-            Assert(entry != null);
-
-            try
-            {
-                _logger?.LogTrace($"[{await GetSessionAsync(cancellation)}] Pending write-lock for entry '{entry.Path}'. Waiting for read-locks to release.");
-
-                // Wait till all read-locks are freed.
-                entry = await WaitForReadLocksReleaseAsync(entry, cancellation);
-
-                // We hold the write lock. No-one can alter the entry except our session is terminated. But this will cause WaitForReadLocksReleaseAsync to throw.
-                Assert(entry != null);
-
-                // We own the write-lock.
-                // All read-locks must be free except for ourself.
-                Assert(entry.WriteLock == session);
-                Assert(entry.ReadLocks.IsEmpty || entry.ReadLocks.Length == 1 && entry.ReadLocks[0] == session);
-
-                if (_logger?.IsEnabled(LogLevel.Trace) ?? false)
-                {
-                    Assert(watch != null);
-                    watch.Stop();
-
-                    _logger?.LogTrace($"[{await GetSessionAsync(cancellation)}] Acquired write-lock for entry '{entry.Path}' in {watch.Elapsed.TotalSeconds}sec.");
-                }
-
-                return entry;
-            }
-            catch
-            {
-                try
-                {
-                    await InternalReleaseWriteLockAsync(entry);
-                }
-                catch (SessionTerminatedException) { throw; }
-                catch
-                {
-                    await DisposeAsync();
-                    throw;
-                }
-
-                throw;
-            }
-        }
-
-        private async Task<IStoredEntry> InternalReleaseWriteLockAsync(IStoredEntry entry)
-        {
-            var cancellation = _disposeHelper.DisposalRequested;
-
-            IStoredEntry start, desired;
-
-            var session = await GetSessionAsync(cancellation);
-
-            if (entry != null)
-            {
-                _logger?.LogTrace($"[{await GetSessionAsync(cancellation)}] Releasing write-lock for entry '{entry.Path}'.");
-            }
-
-            CacheEntry cacheEntry;
-
-            do
-            {
-                start = entry;
-
-                // The entry was deleted.
-                if (start == null)
-                {
-                    NotifyWriteLockReleasedAsync(entry.Path, cancellation).HandleExceptions(_logger);
-                    _logger?.LogTrace($"[{await GetSessionAsync(cancellation)}] Released write-lock for entry '{entry.Path}'.");
-                    return start;
-                }
-
-                // The session does not own the write lock.
-                if (start.WriteLock != session)
-                {
-                    return start;
-                }
-
-                // As we did not invalidate the cache entry on write-lock acquirement, we have to update the cache.
-                // We must not update the cache, before we released the write lock, because it is not guaranteed, 
-                // that the write lock release is successful on the first attempt.
-                // We read the cache-entry before we release the write lock, to get the cache entry version and
-                // afterwards try to update the cache if no other session invalidated our cache entry in the meantime.
-                cacheEntry = _cache.GetEntry(entry.Path);
-                desired = _storedEntryManager.ReleaseWriteLock(start, session);
-                entry = await _storage.UpdateEntryAsync(desired, start, cancellation);
-            }
-            while (entry != start);
-
-            entry = desired;
-            Assert(entry != null);
-            _cache.UpdateEntry(cacheEntry, entry);
-
-            if (entry != null)
-            {
-                NotifyWriteLockReleasedAsync(entry.Path, cancellation).HandleExceptions(_logger);
-                _logger?.LogTrace($"[{await GetSessionAsync(cancellation)}] Released write-lock for entry '{entry.Path}'.");
-            }
-
-            return desired;
-        }
-
-        private async Task<IStoredEntry> AcquireReadLockAsync(CoordinationEntryPath path, CancellationToken cancellation)
-        {
-            // Enter local write lock, as we are modifying the stored entry with entering the read-lock
-            await AcquireLocalLockAsync(path, cancellation);
-            try
-            {
-                // Freshly load the entry from the database.
-                var entry = await _storage.GetEntryAsync(path, cancellation);
-
-                // Enter global read lock.
-                return await InternalAcquireReadLockAsync(entry, cancellation);
-            }
-            finally
-            {
-                await ReleaseLocalLockAsync(path, cancellation);
-            }
-        }
-
-        private Task<IStoredEntry> ReleaseReadLockAsync(CoordinationEntryPath path, CancellationToken cancellation)
-        {
-            return ReleaseReadLockAsync(path).WithCancellation(cancellation);
-        }
-
-        private async Task<IStoredEntry> InternalAcquireReadLockAsync(IStoredEntry entry, CancellationToken cancellation)
-        {
-            Stopwatch watch = null;
-            if (_logger?.IsEnabled(LogLevel.Trace) ?? false)
-            {
-                watch = new Stopwatch();
-                watch.Start();
-            }
-
-            IStoredEntry start, desired;
-
-            var session = await GetSessionAsync(cancellation);
-
-            if (entry != null)
-            {
-                _logger?.LogTrace($"[{await GetSessionAsync(cancellation)}] Acquiring read-lock for entry '{entry.Path}'.");
-            }
-
-            do
-            {
-                start = await WaitForWriteLockReleaseAsync(entry, allowWriteLock: true, cancellation);
-
-                // The entry was deleted (concurrently).
-                if (start == null)
-                {
-                    return null;
-                }
-
-                Assert(start.WriteLock == null || start.WriteLock == session);
-
-                desired = _storedEntryManager.AcquireReadLock(start, session);
-                entry = await _storage.UpdateEntryAsync(desired, start, cancellation);
-            }
-            while (start != entry);
-
-            entry = desired;
-            Assert(entry != null);
-            Assert(entry.ReadLocks.Contains(session));
-
-            try
-            {
-                if (_logger?.IsEnabled(LogLevel.Trace) ?? false)
-                {
-                    Assert(watch != null);
-                    watch.Stop();
-
-                    _logger?.LogTrace($"[{await GetSessionAsync(cancellation)}] Acquired read-lock for entry '{entry.Path.EscapedPath}' in {watch.ElapsedMilliseconds}ms.");
-                }
-
-                return entry;
-            }
-            catch
-            {
-                Assert(entry != null);
-
-                // Release global read lock on failure.
-                await ReleaseReadLockAsync(entry.Path);
-                throw;
-            }
-        }
-
-        private async Task<IStoredEntry> ReleaseReadLockAsync(CoordinationEntryPath path)
-        {
-            try
-            {
-                // Enter local write lock, as we are modifying the stored entry with entering the read-lock
-                await AcquireLocalLockAsync(path, cancellation: default);
-                try
-                {
-                    // Freshly load the entry from the database.
-                    var entry = await _storage.GetEntryAsync(path, cancellation: default);
-
-                    // Release global read lock.
-                    return await InternalReleaseReadLockAsync(entry);
-                }
-                finally
-                {
-                    await ReleaseLocalLockAsync(path, cancellation: default);
-                }
-            }
-            catch
-            {
-                await DisposeAsync();
-                throw;
-            }
-        }
-
-        private async Task<IStoredEntry> InternalReleaseReadLockAsync(IStoredEntry entry)
-        {
-            Assert(entry != null);
-
-            var cancellation = _disposeHelper.DisposalRequested;
-
-            IStoredEntry start, desired;
-
-            var session = await GetSessionAsync(cancellation);
-
-            if (entry != null)
-            {
-                _logger?.LogTrace($"[{await GetSessionAsync(cancellation)}] Releasing read-lock for entry '{entry.Path}'.");
-            }
-
-            do
-            {
-                start = entry;
-
-                // The entry was deleted (concurrently).
-                if (start == null || !start.ReadLocks.Contains(session))
-                {
-                    return null;
-                }
-
-                desired = _storedEntryManager.ReleaseReadLock(start, session);
-
-                entry = await _storage.UpdateEntryAsync(desired, start, cancellation);
-            }
-            while (start != entry);
-
-            Assert(entry != null);
-
-            NotifyReadLockReleasedAsync(entry.Path, cancellation).HandleExceptions(_logger);
-            _logger?.LogTrace($"[{await GetSessionAsync(cancellation)}] Released read-lock for entry '{entry.Path}'.");
-
-            return desired;
-        }
-
-        // WaitForWriteLockReleaseAsync updates the session in order that there is enough time
-        // to complete the write operation, without the session to terminate.
-        private async Task<IStoredEntry> WaitForWriteLockReleaseAsync(IStoredEntry entry, bool allowWriteLock, CancellationToken cancellation)
-        {
-            if (entry != null)
-            {
-                _logger?.LogTrace($"[{await GetSessionAsync()}] Waiting for write-lock release for entry '{entry.Path}'.");
-            }
-
-            var session = await GetSessionAsync(cancellation);
-
-            // The entry was deleted (concurrently).
-            while (entry != null)
-            {
-                var writeLock = entry.WriteLock;
-
-                if (writeLock == null)
-                {
-                    return entry;
-                }
-
-                // If (entry.WriteLock == session) we MUST wait till the lock is released
-                // and acquired again in order that no concurrency conflicts may occur.
-                // Because of the case that we may keep a write and read lock at the same time, this does hold true only if we aquire a write lock.
-                // In this case, we could also allow this but had to synchronize the write operations locally.
-                if (allowWriteLock && writeLock == session)
-                {
-                    return entry;
-                }
-
-                var path = entry.Path;
-
-                async Task<bool> Predicate(CancellationToken c)
-                {
-                    entry = await _storage.GetEntryAsync(path, c);
-                    return entry == null || entry.WriteLock == null;
-                }
-
-                Task Release(CancellationToken c)
-                {
-                    return _writeLockWaitDirectory.WaitForNotificationAsync(((Session)writeLock, path), c);
-                }
-
-                var combinedCancellationSource = CancellationTokenSource.CreateLinkedTokenSource(cancellation);
-
-                try
-
-                {
-                    var lockRelease = SpinWaitAsync(Predicate, Release, combinedCancellationSource.Token);
-                    var sessionTermination = _sessionManager.WaitForTerminationAsync((Session)writeLock);
-                    var completed = await (Task.WhenAny(sessionTermination, lockRelease).WithCancellation(cancellation));
-
-                    if (completed == sessionTermination)
-                    {
-                        await CleanupLocksOnSessionTermination(path, (Session)writeLock, cancellation);
-                        entry = await _storage.GetEntryAsync(path, cancellation);
-                    }
-                }
-                finally
-                {
-                    combinedCancellationSource.Cancel();
-                }
-            }
-
-            return null;
-        }
-
-        private async Task<IStoredEntry> WaitForReadLocksReleaseAsync(IStoredEntry entry, CancellationToken cancellation)
-        {
-            Assert(entry != null);
-
-            if (entry != null)
-            {
-                _logger?.LogTrace($"[{await GetSessionAsync()}] Waiting for read-locks release for entry '{entry.Path}'.");
-            }
-
-            IEnumerable<Session> readLocks = entry.ReadLocks;
-
-            // Exclude our own read-lock (if present)
-            var session = await GetSessionAsync(cancellation);
-            readLocks = readLocks.Where(readLock => readLock != session);
-
-            // Send a message to each of 'readLocks' to ask for lock release and await the end of the session or the lock release, whichever occurs first.
-            await Task.WhenAll(readLocks.Select(readLock => WaitForReadLockRelease(entry.Path, readLock, cancellation)));
-
-            // The unlock of the 'readLocks' will alter the db, so we have to read the entry again.
-            entry = await _storage.GetEntryAsync(entry.Path, cancellation);
-
-            // We are holding the write-lock => The entry must not be deleted concurrently.
-            if (entry == null || entry.ReadLocks.Length != 0 && (entry.ReadLocks.Length > 1 || entry.ReadLocks.First() != session))
-            {
-                throw new SessionTerminatedException();
-            }
-
-            return entry;
-        }
-
-        private async Task CleanupLocksOnSessionTermination(CoordinationEntryPath path, Session session, CancellationToken cancellation)
-        {
-#if DEBUG
-            var isTerminated = !await _sessionManager.IsAliveAsync(session, cancellation);
-
-            Assert(isTerminated);
-#endif
-
-            var localSession = await GetSessionAsync(cancellation);
-
-            // We waited for ourself to terminate => We are terminated now.
-            if (session == localSession)
-            {
-                throw new SessionTerminatedException();
-            }
-
-            IStoredEntry entry = await _storage.GetEntryAsync(path, cancellation),
-                         start,
-                         desired;
-
-            do
-            {
-                cancellation.ThrowIfCancellationRequested();
-
-                start = entry;
-
-                if (start == null)
-                {
-                    return;
-                }
-
-                desired = start;
-
-                if (entry.WriteLock == session)
-                {
-                    desired = _storedEntryManager.ReleaseWriteLock(desired, session);
-                }
-
-                if (entry.ReadLocks.Contains(session))
-                {
-                    desired = _storedEntryManager.ReleaseReadLock(desired, session);
-                }
-
-                if (AreVersionEqual(start, desired))
-                {
-                    return;
-                }
-
-                entry = await _storage.UpdateEntryAsync(desired, start, cancellation);
-            }
-            while (start != entry);
-        }
-
-        private async Task WaitForReadLockRelease(CoordinationEntryPath path, Session session, CancellationToken cancellation)
-        {
-            var entry = await _storage.GetEntryAsync(path, cancellation);
-
-            if (entry == null || !entry.ReadLocks.Contains(session))
-            {
-                return;
-            }
-
-            if (!await _sessionManager.IsAliveAsync(session, cancellation))
-            {
-                await CleanupLocksOnSessionTermination(path, session, cancellation);
-                return;
-            }
-
-            async Task<bool> Predicate(CancellationToken c)
-            {
-                entry = await _storage.GetEntryAsync(path, c);
-
-                if (entry != null && entry.ReadLocks.Contains(session))
-                {
-                    InvalidateCacheEntryAsync(path, session, c).HandleExceptions(_logger);
-                    return false;
-                }
-
-                return true;
-            }
-
-            Task Release(CancellationToken c)
-            {
-                return _readLockWaitDirectory.WaitForNotificationAsync((session, path), c);
-            }
-
-            var combinedCancellationSource = CancellationTokenSource.CreateLinkedTokenSource(cancellation);
-
-            try
-            {
-                var lockRelease = SpinWaitAsync(Predicate, Release, combinedCancellationSource.Token);
-                var sessionTermination = _sessionManager.WaitForTerminationAsync(session);
-                var completed = await (Task.WhenAny(sessionTermination, lockRelease).WithCancellation(cancellation));
-
-                if (completed == sessionTermination)
-                {
-                    await CleanupLocksOnSessionTermination(path, session, cancellation);
-                    entry = await _storage.GetEntryAsync(path, cancellation);
-                }
-            }
-            finally
-            {
-                combinedCancellationSource.Cancel();
-            }
-
-        }
-
-        private async Task SpinWaitAsync(Func<CancellationToken, Task<bool>> predicate, Func<CancellationToken, Task> release, CancellationToken cancellation)
-        {
-            var timeToWait = new TimeSpan(200 * TimeSpan.TicksPerMillisecond);
-            var timeToWaitMax = new TimeSpan(12800 * TimeSpan.TicksPerMillisecond);
-
-            cancellation.ThrowIfCancellationRequested();
-            var releaseX = release(cancellation); // TODO: Rename
-
-            while (!await predicate(cancellation))
-            {
-                cancellation.ThrowIfCancellationRequested();
-
-                var delay = Task.Delay(timeToWait, cancellation);
-
-                var completed = await Task.WhenAny(delay, releaseX);
-
-                if (completed == releaseX)
-                {
-                    return;
-                }
-
-                if (timeToWait < timeToWaitMax)
-                    timeToWait = new TimeSpan(timeToWait.Ticks * 2);
-            }
-        }
-
-        #endregion
 
         private async Task UpdateEntryAsync(IStoredEntry value, IStoredEntry comparand, CancellationToken cancellation)
         {
